+++ R/app_server.R
@@ -30,18 +30,18 @@ app_server <- function(input, output, session) {
   mod_import_cohort_atlas_server("mod_import_cohort_atlas", r_connection, r_cohorts)
   mod_import_cohort_endpoints_server("mod_import_cohort_endpoint", r_connection, r_cohorts)
 
-  mod_cohorts_table_server("mod_cohorts_table_operate", r_cohorts)
-  mod_operate_cohorts_server("mod_operate_cohorts", r_cohorts)
+   mod_cohorts_table_server("mod_cohorts_table_operate", r_cohorts)
+   mod_operate_cohorts_server("mod_operate_cohorts", r_cohorts)
 
-  mod_cohorts_table_server("mod_cohorts_table_compare", r_cohorts, table_editing = FALSE)
-  mod_compare_cohorts_server("mod_compare_cohorts", r_cohorts)
+   mod_cohorts_table_server("mod_cohorts_table_compare", r_cohorts, table_editing = FALSE)
+   mod_compare_cohorts_server("mod_compare_cohorts", r_cohorts)
 
-  mod_cohorts_table_server("mod_cohorts_table_phewas", r_cohorts, table_editing = FALSE)
-  mod_phewas_server("mod_phewas", r_connection, r_cohorts)
-  mod_cohorts_table_server("mod_cohorts_table_gwas", r_cohorts, table_editing = FALSE)
-  mod_gwas_server("mod_gwas", r_connection, r_cohorts)
+   mod_cohorts_table_server("mod_cohorts_table_phewas", r_cohorts, table_editing = FALSE)
+   mod_phewas_server("mod_phewas", r_connection, r_cohorts)
+   mod_cohorts_table_server("mod_cohorts_table_gwas", r_cohorts, table_editing = FALSE)
+   mod_gwas_server("mod_gwas", r_connection, r_cohorts)
 
-
+  #
   # info bubbles ---------------------------------------------
   mod_info_box_server("info_importcohorts", "Import Cohorts", "info_importcohorts.md")
   mod_info_box_server("info_operatecohorts", "Operate Cohorts", "info_operatecohorts.md")