+++ R/fct_config_FGpheWAS.R
@@ -13,6 +13,8 @@
 configFGpheWAS <- function() {
 
 
+
+
   # testing
   if (get_golem_config("enviroment") == "no_connection") {
     return(
@@ -31,7 +33,38 @@ configFGpheWAS <- function() {
     # authenticate
     bigrquery::bq_auth(path = get_golem_config("GCP_SERVICE_KEY"))
 
-    testing_phewas_schema <- "phewas_dummy_1k"
+    # CDMTools config
+    connection_details <- DatabaseConnector::createConnectionDetails(
+      dbms = get_golem_config("CDMTOOLS_dbms"),
+      bq_dbi_project = get_golem_config("GCP_PROJECT_ID"),
+      bq_dbi_billing = bq_dbi_billing
+    )
+
+    connection_settings_n <- list(
+      #
+      sandbox_tools_r6 = FGpheWAS::createConnectionSettings(
+        name = "r6",
+        connection_details = connection_details,
+        phewas_schema = "sandbox_tools_r6",
+        endpoint_cohorts_table = "endpoint_cohorts_dummy50k_v1",
+        code_counts_table = "code_counts_dummy50k_v1",
+        df9_flag = FALSE,
+        codes_info_schema = "medical_codes",
+        fg_codes_info_table = "fg_codes_info_v1"
+      ),
+      #
+      sandbox_tools_r10 = FGpheWAS::createConnectionSettings(
+        name = "r10",
+        connection_details = connection_details,
+        phewas_schema = "sandbox_tools_r10",
+        endpoint_cohorts_table = "endpoint_cohorts_r10_v1",
+        code_counts_table = "code_counts_r10_v1",
+        df9_flag = FALSE,
+        codes_info_schema = "medical_codes",
+        fg_codes_info_table = "fg_codes_info_v1"
+      )
+    )
+
   }
 
   # in development sandbox
@@ -48,20 +81,7 @@ configFGpheWAS <- function() {
   }
 
 
-  # CDMTools config
-  connection_details <- DatabaseConnector::createConnectionDetails(
-    dbms = get_golem_config("CDMTOOLS_dbms"),
-    bq_dbi_project = get_golem_config("GCP_PROJECT_ID"),
-    bq_dbi_billing = bq_dbi_billing
-  )
-
-
-
-  connection_settings <- FGpheWAS::createConnectionSettings(
-    connection_details = connection_details,
-    phewas_schema =testing_phewas_schema
-  )
 
   # print(cdm_webapi_conn)
-  return(connection_settings)
+  return(connection_settings_n)
 }