+++ R/mod_phewas.R
@@ -13,6 +13,10 @@ mod_phewas_ui <- function(id){
     shinyjs::useShinyjs(),
     shinyWidgets::useSweetAlert(),
     #
+    shiny::tags$h4("Select Data Freeze"),
+    # TOFIX: couldnt make to updateSelectInput with in the module, this is plan B
+    shiny::uiOutput(ns("updated_selectinput")),
+    #
     shiny::tags$h4("Cohorts to compare"),
     shiny::fluidRow(
       shiny::column(6,shiny::tags$h5("Cases-cohort"),
@@ -143,6 +147,26 @@ mod_phewas_server <- function(id, r_connection, r_cohorts){
       analysis_settings = NULL
     )
 
+    #
+    # creates picker in output$updated_selectinput
+    #
+    output$updated_selectinput <- shiny::renderUI({
+      shiny::req(r_connection$phewas_conn)
+
+      df_choices <- names(r_connection$phewas_conn)
+
+      if (length(df_choices)==0) {
+        shiny::selectInput(ns("database_picker"), "Select CDM database:", choices = NULL)
+      } else {
+        shiny::selectInput(ns("database_picker"), "Select CDM database:",
+                           choices = df_choices,
+                           selected = df_choices[1]
+        )
+      }
+    })
+
+
+
     shiny::observe({
       shiny::req(r_cohorts$cohortData)
 
@@ -186,7 +210,7 @@ mod_phewas_server <- function(id, r_connection, r_cohorts){
         CohortOperationsShinyApp::sweetAlert_spinner("Assessing patients in cases and controls")
 
         cohorts_settings <- FGpheWAS::createCohortsSettings(
-          connection_settings = r_connection$phewas_conn,
+          connection_settings = r_connection$phewas_conn[[input$database_picker]],
           cases_cohort_source = cases_cohort %>% dplyr::distinct(COHORT_SOURCE) %>% dplyr::pull(COHORT_SOURCE),
           cases_cohort_name = cases_cohort %>% dplyr::distinct(COHORT_NAME) %>% dplyr::pull(COHORT_NAME),
           cases_id_list = cases_id_list,
@@ -305,7 +329,7 @@ mod_phewas_server <- function(id, r_connection, r_cohorts){
         #
         CohortOperationsShinyApp::sweetAlert_spinner("Fetching cohort counts")
         case_controls_counts <- FGpheWAS::getCodeCounts(
-          connection_settings = r_connection$phewas_conn,
+          connection_settings = r_connection$phewas_conn[[input$database_picker]],
           cohorts_settings = r_phewas$cohorts_settings,
           analysis_settings = r_phewas$analysis_settings
         )